Updated version

+++ src/os/type/Cookie.java
@@ -1,12 +1,11 @@
 package os.type;
 
-
 public class Cookie {
 
 	private String key;
 	private String value;
-	private int age;
-	
+	private long age;
+
 	public Cookie(String key, String value) {
 
 		this.key = key;
@@ -14,32 +13,32 @@ public class Cookie {
 	}
 
 	public String getKey() {
-	
+
 		return key;
 	}
-	
+
 	public void setKey(String key) {
-	
+
 		this.key = key;
 	}
-	
+
 	public String getValue() {
-	
+
 		return value;
 	}
-	
+
 	public void setValue(String value) {
-	
+
 		this.value = value;
 	}
-	
-	public int getAge() {
-	
+
+	public long getAge() {
+
 		return age;
 	}
-	
-	public void setAge(int age) {
-	
+
+	public void setAge(long age) {
+
 		this.age = age;
 	}
 }
